Implement a very simple (but slow) mtie calculation
Implement the library function mtie(), which does a brute-force
calculation of mtie over a set of evenly spaced samples.

The function is limited to only accept inputs of 50,000 samples
or fewer.  Experimentation indicates that inputs of that size
complete in a reaonsable amount of time (less than a minute),
but bigger input sizes take too long to process.