Ignore invalid input
Graceful ignore invalid input, outputting to standard error
that the line is ignored, and outputting the contents of that line.

It is TBD whether that is a reasonable design choice.  Possibly
a better design is to explicitly add support for "comments" in the
input data file, and allow that, but disallow anything else.