Choose the MTIE algorithm based on the number of samples
Dynamically choose between the "fast" (but incomplete) or the
"complete" (but slow) MTIE algorithms, based upon the size
of the TIE input data sample.

Use the complete MTIE algorithm if the TIE input data size
is small enough that the calculation can happen in a reasonable
amount of time.  Use the fast MTIE algorithm otherwise.

The threshold for choosing between the algorithms is set to
100,000 samples of TIE, based on experimentation on my PC.